+++ packages/vrender-core/src/graphic/graphic.ts
@@ -27,6 +27,7 @@ import { Node } from './node-tree';
 import type {
   IAnimate,
   IAnimateTarget,
+  IFullThemeSpec,
   IGlyphGraphicAttribute,
   IGroup,
   ILayer,
@@ -34,7 +35,8 @@ import type {
   IShadowRoot,
   IStage,
   IStep,
-  ISubAnimate
+  ISubAnimate,
+  ITheme
 } from '../interface';
 import { EventTarget, CustomEvent } from '../event';
 import { DefaultTransform } from './config';
@@ -262,6 +264,8 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   declare finalAttrs?: T;
 
   declare pathProxy?: ICustomPath2D;
+  // 依附于某个theme，如果该节点不存在parent，那么这个Theme就作为节点的Theme，避免添加到节点前计算属性
+  declare attachedThemeGraphic?: IGraphic;
 
   constructor(params: T = {} as T) {
     super();
@@ -313,7 +317,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
 
     application.graphicService.beforeUpdateAABBBounds(this, this.stage, true, this._AABBBounds);
     const bounds = this.doUpdateAABBBounds(full);
-    this.addUpdateLayoutTag();
+    // this.addUpdateLayoutTag();
     application.graphicService.afterUpdateAABBBounds(this, this.stage, this._AABBBounds, this, true);
     return bounds;
   }
@@ -1373,9 +1377,11 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
         ResourceLoader.GetImage(image, this);
         this.backgroundImg = this.backgroundImg || background;
       }
-    } else {
+    } else if (isObject(image)) {
       (cache.state = 'success'), (cache.data = image);
       this.backgroundImg = this.backgroundImg || background;
+    } else {
+      cache.state = 'fail';
     }
   }