+++ packages/vrender-core/src/graphic/node-tree.ts
@@ -414,11 +414,13 @@ export class Node extends EventEmitter<any, any> implements INode {
     }
     let child = this._firstChild;
     while (child) {
+      const next = child._next;
       child.parent = null;
       //   child._rootNode = child;
       child._prev = null;
       child._next = null;
       child = child._next;
+      child = next;
     }
 
     this._firstChild = null;