+++ packages/vrender-components/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log - @visactor/vrender-components
 
-This log was last generated on Wed, 24 Jan 2024 13:11:27 GMT and should not be manually modified.
+This log was last generated on Thu, 01 Feb 2024 09:26:17 GMT and should not be manually modified.
+
+## 0.17.20
+Thu, 01 Feb 2024 09:26:17 GMT
+
+### Updates
+
+- feat: unify richtext config
+- feat: support richtext in legend components
+- feat: tag support textAlwaysCenter, closed #915
 
 ## 0.17.19
 Wed, 24 Jan 2024 13:11:27 GMT