+++ packages/vrender-kits/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @visactor/vrender-kits
 
-This log was last generated on Wed, 24 Jan 2024 13:11:27 GMT and should not be manually modified.
+This log was last generated on Thu, 01 Feb 2024 09:26:17 GMT and should not be manually modified.
+
+## 0.17.20
+Thu, 01 Feb 2024 09:26:17 GMT
+
+_Version update only_
 
 ## 0.17.19
 Wed, 24 Jan 2024 13:11:27 GMT